Updated Task 1 Algorithm, Making it more efficient
Before, The task 1 Train Journey Display algorithm had an IF statement in the FOR loop to check if the tickets available or not. If tickets were not available it would show that the Train for that hour is closed. But knowing that this is the first time the Train Journey Display is being shown, and there had not been any bookings. We can remove the IF statements from the task 1 because there is no way that a train is going to be closed when the program has just started.